requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
pytorch-lightning==2.0.6
hdbscan==0.8.33
umap-learn==0.5.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
jupyterlab==4.0.3
ipykernel==6.25.0
ipywidgets==8.0.7
mlflow==2.5.0
pyyaml==6.0
tqdm==4.65.0
openpyxl==3.1.2
shap==0.42.1
yellowbrick==1.5
prince==0.7.1
statsmodels==0.14.0
pingouin==0.5.3
fastapi==0.100.0
uvicorn[standard]==0.23.1
pydantic==2.1.1
python-dotenv==1.0.0
requests==2.31.0
